+++ apps/dashboard/src/app/(dashboard)/(chain)/[chain_id]/[contractAddress]/modules/components/moduleContractInfo.ts
