+++ apps/dashboard/src/app/(dashboard)/(chain)/[chain_id]/[contractAddress]/settings/components/detected-state.tsx
